chore(semantic-release.yml): replace Node.js setup with Bun setup
The changes were made to switch from Node.js to Bun for setting up the
environment and running commands in the GitHub Actions workflow. Bun is
a faster, zero-config build tool for modern JavaScript and TypeScript
codebases. The use of the master branch of the devmasx/merge-branch
action ensures that the latest updates and bug fixes are always used.