fix(package.json): pin semantic-release related packages to specific versions
The semantic-release related packages are now pinned to specific
versions instead of using the caret (^) to allow minor updates. This
change ensures that the build process is more predictable and stable, as
it will always use the same versions of these packages. This can help
prevent unexpected build failures due to updates in these packages.