feat: add support for querying strains by gene in the GraphQL schema and mock data
The changes introduce a new query type `listStrainsWithGene` in the GraphQL schema to allow querying strains by gene. Additionally, mock data and resolver functions are added to support mocking this query for testing purposes. This enhancement improves the flexibility of the system by enabling querying strains based on a specific gene.