feat(query.graphql): add new queries listPhenotypeEnvironments and listPhenotypeAssays
The new queries listPhenotypeEnvironments and listPhenotypeAssays have been added to the Query type in the GraphQL schema. These queries allow users to search for phenotype environments and phenotype assays respectively, by providing a search string. This enhancement expands the functionality of the API by enabling users to retrieve specific information related to phenotype environments and assays.