feat: generate dist
This commit introduces the apollo client helpers and fragment types.
The apollo client helpers file contains type definitions for the
Apollo Client cache policies. The fragment file contains type
definitions for the possible types of the BasePublication and Stock
interfaces. These files are necessary for the Apollo Client to
correctly cache and retrieve data from the GraphQL API.

This commit adds the generated typescript definitions for the
graphql schema. This allows the frontend to use the generated
types for the graphql schema.

This commit adds autogenerated apollo graphql queries and mutations.
This allows the frontend to easily query and mutate data from the
backend.

This commit introduces a new file, `dist/mocks.ts`, which contains mock
implementations for all GraphQL queries and mutations defined in the
application. These mocks are intended to be used with MSW (Mock Service
Worker) to facilitate testing and development by providing predictable
responses without relying on a live backend.

The generated mocks cover a wide range of queries and mutations, including
those for authentication, content management, order processing, strain and
plasmid management, user management, and data retrieval for various
entities like organisms, genes, publications, and phenotypes.

The addition of these mocks enables developers to:

-   Test UI components in isolation without a backend.
-   Develop new features against a stable API contract.
-   Simulate different scenarios and edge cases.
-   Improve the overall testability of the application.

The mocks are generated based on the GraphQL schema and are designed to be
easily customizable to fit specific testing needs.